```text feat: Update image.service.js to remove triggerWebhook call
The code changes in this commit remove the triggerWebhook call in the updateRequestStatus function of image.service.js. This call is commented out to prevent the webhook request from being sent upon completion of image processing. This change aligns with the recent user commits that added the webhookUrl configuration and indicates that the webhook request will be handled separately.